[OPIK-7105] [FE] refactor: retire the workspace-version machinery
Nothing in the frontend asks which experience a workspace is on any more.

Removed: the version gate (the entry point now mounts V2App directly), the
WorkspaceVersion type and its store field, setters and selectors, the resolution
library and its test, and the query hook that returned a literal without calling
the endpoint the backend already removed.

WorkspaceVersionResolver is deleted, not refactored. Everything it did was
version machinery -- writing "v2" to the store, mirroring it into a localStorage
cache nothing read, and a reload path unreachable because the value was constant.
With the store field gone it had no remaining effect, so it is unwired from the
V2 workspace guard and removed.

PairRouteVersionGuard did have one live effect and is replaced rather than
deleted. Pair routes render outside WorkspaceGuard, so nothing else lifts the
active workspace out of ?workspace=. PairRouteWorkspaceGuard keeps exactly that
and drops the rest.

That also fixes a pre-existing defect. The old guard held on
`!active || !apiVersion || mismatch`, and the version query was disabled without
a workspace name -- so a pairing URL with no ?workspace= parameter sat on a
loader forever. The new guard waits only while a workspace named in the URL has
yet to be applied; with none in the URL there is nothing to wait for, so the page
renders and reports the missing parameter itself.

getWorkspaceNameFromUrl was the one function in the resolution library with real
value; it moves to lib/utils next to buildFullBaseUrl. navigateToWorkspaceRoot
lost its last caller when the user-menu toggle went, and is dropped.

Also removes the no-op version-override seed from E2E global setup, whose doc
comment still described determination via an endpoint that no longer exists, and
three residual TOGGLE_FORCE_WORKSPACE_VERSION references in the OpenAPI examples
override, the self-host migration page and both dev-runner scripts.

Retiring workspaceVersion.test.ts is deliberate. It guarded a real incident -- a
stale locally-stored version breaking SDK pairing -- by asserting the resolver
ignored legacy local state. With every read of that state gone, there is nothing
left for it to protect.

Cost Intelligence was already decoupled from this machinery in an earlier commit,
so removing the writers here cannot affect it. Verified: no workspace-version
reference remains anywhere under src, and the CI-equivalent build with the
private ai-spend overlay staged passes.
