+++ apps/opik-documentation/documentation/fern/docs-v2/self-host/v1_to_v2_migration.mdx
@@ -10,7 +10,7 @@ title: Migrating to Opik 2.0
 Opik 2.0 organizes everything around **projects**. Data created after upgrading is automatically project-scoped. Data created on Opik 1.x — datasets, prompts, experiments, optimizations, automation rules, and alerts — requires six background jobs to assign each item to the right project. This page walks you through running those jobs.
 
 <Warning>
-  **Upgrading from Opik 1.x with existing data?** Recent Opik images default `TOGGLE_FORCE_WORKSPACE_VERSION` to `version_2`, which loads the Opik 2.0 UI globally regardless of your existing entities. 1.x entities that have not yet been assigned to a project (no `project_id`) will not be visible in the 2.0 UI until you run the migration jobs below.
+  **Upgrading from Opik 1.x with existing data?** Recent Opik images ship the Opik 2.0 UI as the only experience, regardless of your existing entities. 1.x entities that have not yet been assigned to a project (no `project_id`) will not be visible in the 2.0 UI until you run the migration jobs below.
 </Warning>
 
 ## Before you start
