+++ apps/opik-documentation/documentation/fern/openapi/ai_examples_override.yml
@@ -3609,7 +3609,6 @@ paths:
               customllmProviderEnabled: true
               ollamaProviderEnabled: false
               collaboratorsTabEnabled: true
-              forceWorkspaceVersion: version_2
   /v1/private/experiments/batch:
     patch:
       x-fern-examples:
