+++ apps/opik-frontend/src/index.tsx
@@ -3,10 +3,13 @@ import * as Sentry from "@sentry/react";
 
 import "tailwindcss/tailwind.css";
 
-import WorkspaceVersionGate from "@/WorkspaceVersionGate";
+import React, { Suspense } from "react";
 import usePluginsStore from "@/store/PluginsStore";
+import Loader from "@/shared/Loader/Loader";
 import { APP_VERSION } from "@/constants/app";
 import { runLocalStorageMigrations } from "@/lib/ls-migrations";
+
+const V2App = React.lazy(() => import("@/v2/App"));
 
 import "./main.scss";
 import { IS_SENTRY_ENABLED, SENTRY_DSN, SENTRY_MODE } from "@/config";
@@ -31,7 +34,11 @@ if (IS_SENTRY_ENABLED) {
 async function bootstrap() {
   await usePluginsStore.getState().setupPlugins();
   runLocalStorageMigrations();
-  root.render(<WorkspaceVersionGate />);
+  root.render(
+    <Suspense fallback={<Loader />}>
+      <V2App />
+    </Suspense>,
+  );
 }
 
 void bootstrap();
