+++ apps/opik-frontend/src/shared/OAuthConsentPage/createOAuthConsentRoute.ts
@@ -3,9 +3,8 @@ import { AnyRoute, createRoute } from "@tanstack/react-router";
 
 const OAuthConsentPage = lazy(() => import("./OAuthConsentPage"));
 
-// Single source of truth for the pre-workspace MCP OAuth consent route. Both the v1 and v2
-// routers mount it at the app root because WorkspaceVersionGate picks the router before any
-// workspace is resolved and the consent URL is workspace-version-agnostic.
+// Single source of truth for the pre-workspace MCP OAuth consent route. The router mounts it
+// at the app root because the consent URL is reached before any workspace is resolved.
 export const createOAuthConsentRoute = <T extends AnyRoute>(rootRoute: T) =>
   createRoute({
     getParentRoute: () => rootRoute,
